requirements.txt
tensorflow==2.9.1
Flask
dvc
pandas
notebook
numpy
matplotlib
seaborn
python-box
pyYAML
tqdm
ensure
joblib
types-PyYAML
scipy>=1.9.1
opencv-python
streamlit


-e .